Update engine.lua
treeCanvas and TwoCanvas did nor work on my hardware. It seems these canvases have no dept buffer and the picture is not right.

Drawing directly to the screen canvas works ok. Hence i implemented a direct screen draw mode. In addition it saves a lot of memory used for the 2 canvases.